Move `WorkerPool` to `execution`
