+++ packages/engine/lib/execution/Cargo.toml
@@ -18,6 +18,7 @@ glob = "0.3.0"
 kdtree = "0.6.0"
 lazy_static = "1.4.0"
 nng = { version = "1.0.1", default-features = false }
+rand = "0.8.3"
 rayon = "1.4.1"
 serde = { version = "1.0.111", features = ["derive"] }
 serde_json = "1.0.59"
